requirements.txt
psutil
pyperclip
Pillow